requirements.txt
claude-code-sdk>=0.1.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
yfinance>=0.2.18
ta>=0.10.2
scipy>=1.11.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
websockets>=11.0.0
pydantic>=2.0.0
loguru>=0.7.0
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2023.3
aiofiles>=23.0.0
psutil>=5.9.0
psycopg2-binary>=2.9.0
finance-tools-mcp>=0.1.0